==> TASK4/0_preprocess_data.ipynb <==
import os
import pickle
import tqdm
import pandas as pd
import selfies as sf

class DATACONFIGS:
    def __init__(self):
        ## input
        self.input_dir = os.path.join(os.pardir, 'data', 'zinc15')
        self.train_data_path = os.path.join(self.input_dir, 'zinc15_train.txt')
        ## output
        self.output_dir = 'outputs_0_preprocess_data'
        assert os.path.exists(self.output_dir)       

dataconfigs = DATACONFIGS()

with open(dataconfigs.train_data_path) as fin:
    lines = fin.readlines()
    
data = [line.rstrip() for line in lines]

print(f'Number of training data (raw): {len(data)}')
data[:5]

def make_selfies_data(smiles_iter):
    selfies_list = []
    
    for smi in tqdm.tqdm(smiles_iter):
        try:
            sel = sf.encoder(smi)
            selfies_list.append(sel)
        except AssertionError:
            pass
        except sf.EncoderError:
            pass
    
    return selfies_list

selfies_list = make_selfies_data(data)

print(f'Number of training data (selfies): {len(selfies_list)}')
selfies_list[:5]

with open(os.path.join(dataconfigs.output_dir, 'selfies.csv'), 'w') as fout:
    for selfies in selfies_list:
        fout.write(f'{selfies}\n')

fragments_list = [list(sf.split_selfies(sel)) for sel in selfies_list]

print(f'maxlen of fragments: {max([len(fragments) for fragments in fragments_list])}')
fragments_list[:2]

with open(os.path.join(dataconfigs.output_dir, 'fragments_list.pkl'), 'wb') as fout:
    pickle.dump(fragments_list, fout)

vocabs = sf.get_alphabet_from_selfies(selfies_list)

vocabs = sorted(vocabs)

print(f'Number of vocabulary(unique fragments): {len(vocabs)}')
vocabs

with open(os.path.join(dataconfigs.output_dir, 'vocabulary.csv'), 'w') as fout:
    for v in vocabs:
        fout.write(f'{v}\n')
