==> TASK4/2_optimize.ipynb <==
# BA optimization

%env CUDA_VISIBLE_DEVICES=0

import os
import sys
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import selfies as sf
import torch
import time
from tqdm import tqdm, trange
from rdkit import RDLogger
RDLogger.DisableLog('rdApp.*')

## Inter-op parallelism
torch.set_num_interop_threads(4)
torch.get_num_interop_threads()
## Intra-op parallelism
torch.set_num_threads(4)
torch.get_num_threads()

class GPUCONFIGS:
    def __init__(self):
        self.use_cuda = torch.cuda.is_available()
        self.device = torch.device('cuda:0' if self.use_cuda else 'cpu')
        if self.use_cuda: torch.cuda.set_device(self.device)
        
gpuconfigs = GPUCONFIGS()
print(torch.cuda.current_device())

REBADD_LIB_PATH = os.path.abspath(os.pardir)
if REBADD_LIB_PATH not in sys.path:
    sys.path = [REBADD_LIB_PATH] + sys.path

from rebadd.stackVAE import StackAugmentedVAE
from rebadd.datautils import GeneratorData
from rebadd.reinforce import REINFORCE_SCST_OFFPOLICY as REINFORCE

# 1. Generator

## 1-1) Data loading for generators

class DATACONFIGS:
    def __init__(self):
        ## input
        self.input_dir = 'outputs_0_preprocess_data'
        self.train_data_path = os.path.join(self.input_dir, 'fragments_list.pkl')
        self.vocab_data_path = os.path.join(self.input_dir, 'vocabulary.csv')
        ## output
        self.output_dir = 'outputs_2_optimize'
        assert os.path.exists(self.output_dir)

dataconfigs = DATACONFIGS()

gen_data = GeneratorData(pickle_data_path=dataconfigs.train_data_path,
                         vocabulary_path=dataconfigs.vocab_data_path,
                         use_cuda=gpuconfigs.use_cuda)

print(f"Number of training samples: {len(gen_data.data)}")
print(f"Number of vocabulary: {len(gen_data.vocabs)}")
print(f"Maximum of seqlen: {gen_data.max_seqlen}")

## 1-2) Pretrained generators

kwargs_generator = {"input_size"         : gen_data.n_characters,
                    "output_size"        : gen_data.n_characters,
                    "max_seqlen"         : 44,
                    "hidden_size"        : 256,
                    "latent_size"        : 64,
                    "n_layers"           : 4,
                    "has_stack"          : True,
                    "stack_width"        : 256,
                    "stack_depth"        : 20,
                    "lr"                 : 1e-4,
                    "use_cuda"           : gpuconfigs.use_cuda,
                    "device"             : gpuconfigs.device,
                    "optimizer_instance" : torch.optim.RMSprop}

my_generator_after = StackAugmentedVAE(**kwargs_generator)

class CKPTCONFIGS:
    def __init__(self):
        self.model_path = os.path.join('outputs_1_pretraining', 'checkpoint.pth')
        
ckptconfigs = CKPTCONFIGS()

my_generator_after.load_model(ckptconfigs.model_path)

# 2. Reinforcement learning

from ReBADD_config import Reward_bcl2_bclxl_bclw as Reward

## 2-1) RL design

my_reward_ft = Reward(use_cuda=gpuconfigs.use_cuda, device=gpuconfigs.device)

ReBADD = REINFORCE(data=gen_data, generator=my_generator_after, reward_ft=my_reward_ft)

## 2-2) Training

model_path = os.path.join(dataconfigs.output_dir, 'checkpoint.pth')
losses_path = os.path.join(dataconfigs.output_dir, 'losses.csv')
plot_path = os.path.join(dataconfigs.output_dir, "plots_reward_losses.pdf")

def SMILES_generate(generator, n_to_generate, gen_data):
    generated = []
    for i in trange(n_to_generate):
        z = generator.sample_latent_vectors()
        sel = generator.evaluate(gen_data, z=z)
        sel = sel.replace(gen_data.start_token, "").replace(gen_data.end_token, "")
        smi = sf.decoder(sel)
        generated.append(smi)
    return generated

def simple_moving_average(previous_values, new_value, ma_window_size=10):
    value_ma = np.sum(previous_values[-(ma_window_size-1):]) + new_value
    value_ma = value_ma/(len(previous_values[-(ma_window_size-1):]) + 1)
    return value_ma

def show_reward_loss(rewards, losses, figsize=(8,2.5), filepath=None):
    fig, ax = plt.subplots(1,2,figsize=figsize)
    
    ax[0].plot(rewards)
    ax[0].set_xlabel('Training iteration')
    ax[0].set_ylabel('Average reward')
    ax[1].plot(rl_losses, 'r')
    ax[1].set_xlabel('Training iteration')
    ax[1].set_ylabel('Loss')
    
    plt.tight_layout()
    if filepath is None:
        plt.show()
    else:
        plt.savefig(filepath)

def write_history(tr_rewards, rl_losses, reconst_losses, entropy_losses, ba_A_list, ba_B_list, ba_C_list, execution_times, filepath):
    df_history = pd.DataFrame({"REWARD":tr_rewards,
                               "LOSS":rl_losses,
                               "LOSS_RECONST":reconst_losses,
                               "LOSS_ENTROPY":entropy_losses,
                               "BA_A":ba_A_list,
                               "BA_B":ba_B_list,
                               "BA_C":ba_C_list,
                               "EXEC_TIME":execution_times})
    df_history.to_csv(filepath, index=False)
    return df_history

# Setting up some parameters for the experiment
n_iterations = 500
checkpoint_stepsize = 10
behavior_stepsize = 20

tr_rewards = [] # rewards when training the model
rl_losses = [] # loss when training the model
reconst_losses = [] # reconstruction loss
entropy_losses = [] # entropy regularization

ba_A_list = []
ba_B_list = []
ba_C_list = []

execution_times = []

## Training RL optimizer
start_time = time.time()

alpha_min = 0.01
alpha_max = 0.1
alpha_inc = (alpha_max - alpha_min) / n_iterations
alpha = alpha_min
for i in range(1, n_iterations+1):
    ## policy gradient
    alpha += alpha_inc
    cur_reward, cur_loss, cur_reconst, cur_entropy = ReBADD.policy_gradient(n_batch=8, grad_clipping=1., alpha=alpha)
    
    ## append - reward
    reward = cur_reward
    tr_rewards.append(reward)
    
    ## append - loss
    rl_loss = cur_loss
    rl_losses.append(rl_loss)
    reconst_loss = cur_reconst
    reconst_losses.append(reconst_loss)
    entropy_loss = cur_entropy
    entropy_losses.append(entropy_loss)
        
    ## execution time check
    end_time = time.time()
    exec_time = end_time - start_time
    execution_times.append(exec_time)
    
    ## print - logs
    print(f"[{i:04d}] rl_loss: {rl_loss:.3f}\treconst_loss: {reconst_loss:.3f}\tentropy_loss: {entropy_loss:.3f}\treward: {reward:.3f}\ttau: {ReBADD.tau:.3f}\t({exec_time:.3f} sec)")
        
    ## checkpoint
    avg_ba_A = 0.
    avg_ba_B = 0.
    avg_ba_C = 0.
    if i % checkpoint_stepsize == 0:
        ## save
        ReBADD.generator.save_model(model_path.replace('.pth', f'.pth.{i:04d}'))
        
        ## plotting
        show_reward_loss(tr_rewards, rl_losses)

        ## sampling check
        smiles_cur = SMILES_generate(ReBADD, n_to_generate=100, gen_data=gen_data)
        print(f'Sample trajectories: {len(smiles_cur)}')
        if len(smiles_cur) > 0:
            for sm in smiles_cur:
                _, ba_A, ba_B, ba_C, score_A, score_B, score_C = my_reward_ft(sm, debug=True)
                print(f"[{ba_A:.3f}, {ba_B:.3f}, {ba_C:.3f}, {score_A:.3f}, {score_B:.3f}, {score_C:.3f}]    {sm}")
                avg_ba_A += ba_A
                avg_ba_B += ba_B
                avg_ba_C += ba_C
            avg_ba_A /= len(smiles_cur)
            avg_ba_B /= len(smiles_cur)
            avg_ba_C /= len(smiles_cur)
    
    ## append - binding affinity values
    ba_A_list.append(avg_ba_A)
    ba_B_list.append(avg_ba_B)
    ba_C_list.append(avg_ba_C)
    
    ## history save
    _ = write_history(tr_rewards, rl_losses, reconst_losses, entropy_losses, ba_A_list, ba_B_list, ba_C_list, execution_times, losses_path)
    
    if i % behavior_stepsize == 0:
        ReBADD.update_behavior()

# 3. Save

ReBADD.generator.save_model(model_path)

show_reward_loss(tr_rewards, rl_losses, filepath=plot_path)

df_history = write_history(tr_rewards, rl_losses, reconst_losses, entropy_losses, ba_A_list, ba_B_list, ba_C_list, execution_times, losses_path)
print(df_history.shape)
print(df_history.head())

def save_smiles(filepath, smiles):
    with open(filepath, 'w') as fout:
        for smi in smiles:
            fout.write(f"{smi}\n")
    print(f"[INFO] {len(smiles)} SMILES were saved in {filepath}")

n_sampling = 30000

smi_after = SMILES_generate(ReBADD, n_sampling, gen_data)

save_smiles(os.path.join(dataconfigs.output_dir, "smi_after.csv"), smi_after)
